[WIP] Optimize Regin: Less clones, more HashMaps, shortcircuit expl
- Avoid cloning the whole graph each time `edge_joins_sccs` is
called. Instead add and remove the edge from the original graph each
time. Remove usually invalidates the other indices, but it should work
if the removed edge is the last one.
- Use a HashMap to store intermediate edges instead of a HashSet. This
allows retreiving the (FromIndex, ToIndex) while keeping (roughly) the same
functionality and performance of the HashSet. This allows using less clones.
- Shortcircuit the explanation: Don't call edge_joins_sccs if the edge
is already a valid assignment (and is thus in the graph) because only
assignments not currently valid can be a part of the explanation.